+++ ChampuMusic/utils/decorators/play.py
@@ -9,6 +9,10 @@ from pyrogram.errors import (
 )
 from pyrogram.types import InlineKeyboardButton, InlineKeyboardMarkup
 
+from config import PLAYLIST_IMG_URL, PRIVATE_BOT_MODE
+from config import SUPPORT_GROUP
+from config import adminlist
+from strings import get_string
 from ChampuMusic import YouTube, app
 from ChampuMusic.misc import SUDOERS
 from ChampuMusic.utils.database import (
@@ -18,11 +22,11 @@ from ChampuMusic.utils.database import (
     get_playmode,
     get_playtype,
     is_active_chat,
+    is_commanddelete_on,
     is_maintenance,
+    is_served_private_chat,
 )
 from ChampuMusic.utils.inline import botplaylist_markup
-from config import PLAYLIST_IMG_URL, SUPPORT_GROUP, adminlist
-from strings import get_string
 
 links = {}
 
@@ -42,7 +46,7 @@ def PlayWrapper(command):
                     ]
                 ]
             )
-            return await message.reply_text(_["general_3"], reply_markup=upl)
+            return await message.reply_text(_["general_4"], reply_markup=upl)
 
         if await is_maintenance() is False:
             if message.from_user.id not in SUDOERS:
@@ -50,12 +54,17 @@ def PlayWrapper(command):
                     text=f"{app.mention} ɪs ᴜɴᴅᴇʀ ᴍᴀɪɴᴛᴇɴᴀɴᴄᴇ, ᴠɪsɪᴛ <a href={SUPPORT_GROUP}>sᴜᴘᴘᴏʀᴛ ᴄʜᴀᴛ</a> ғᴏʀ ᴋɴᴏᴡɪɴɢ ᴛʜᴇ ʀᴇᴀsᴏɴ.",
                     disable_web_page_preview=True,
                 )
-                
-
-        try:
-            await message.delete()
-        except:
-            pass
+        if PRIVATE_BOT_MODE == str(True):
+            if not await is_served_private_chat(message.chat.id):
+                await message.reply_text(
+                    "**ᴘʀɪᴠᴀᴛᴇ ᴍᴜsɪᴄ ʙᴏᴛ**\n\nᴏɴʟʏ ғᴏʀ ᴀᴜᴛʜᴏʀɪᴢᴇᴅ ᴄʜᴀᴛs ғʀᴏᴍ ᴛʜᴇ ᴏᴡɴᴇʀ. ᴀsᴋ ᴍʏ ᴏᴡɴᴇʀ ᴛᴏ ᴀʟʟᴏᴡ ʏᴏᴜʀ ᴄʜᴀᴛ ғɪʀsᴛ."
+                )
+                return await app.leave_chat(message.chat.id)
+        if await is_commanddelete_on(message.chat.id):
+            try:
+                await message.delete()
+            except:
+                pass
 
         audio_telegram = (
             (message.reply_to_message.audio or message.reply_to_message.voice)
@@ -75,13 +84,13 @@ def PlayWrapper(command):
                 buttons = botplaylist_markup(_)
                 return await message.reply_photo(
                     photo=PLAYLIST_IMG_URL,
-                    caption=_["play_18"],
+                    caption=_["playlist_1"],
                     reply_markup=InlineKeyboardMarkup(buttons),
                 )
         if message.command[0][0] == "c":
             chat_id = await get_cmode(message.chat.id)
             if chat_id is None:
-                return await message.reply_text(_["setting_7"])
+                return await message.reply_text(_["setting_12"])
             try:
                 chat = await app.get_chat(chat_id)
             except:
@@ -96,7 +105,7 @@ def PlayWrapper(command):
             if message.from_user.id not in SUDOERS:
                 admins = adminlist.get(message.chat.id)
                 if not admins:
-                    return await message.reply_text(_["admin_13"])
+                    return await message.reply_text(_["admin_18"])
                 else:
                     if message.from_user.id not in admins:
                         return await message.reply_text(_["play_4"])
@@ -109,13 +118,13 @@ def PlayWrapper(command):
                 video = True if message.command[0][1] == "v" else None
         if message.command[0][-1] == "e":
             if not await is_active_chat(chat_id):
-                return await message.reply_text(_["play_16"])
+                return await message.reply_text(_["play_18"])
             fplay = True
         else:
             fplay = None
 
         if not await is_active_chat(chat_id):
-            userbot = await get_assistant(chat_id)
+            userbot = await get_assistant(message.chat.id)
             try:
                 try:
                     get = await app.get_chat_member(chat_id, userbot.id)
@@ -125,11 +134,12 @@ def PlayWrapper(command):
                     get.status == ChatMemberStatus.BANNED
                     or get.status == ChatMemberStatus.RESTRICTED
                 ):
-                    return await message.reply_text(
-                        _["call_2"].format(
-                            app.mention, userbot.id, userbot.name, userbot.username
+                    try:
+                        await app.unban_chat_member(chat_id, userbot.id)
+                    except:
+                        return await message.reply_text(
+                            text=_["call_2"].format(userbot.username, userbot.id),
                         )
-                    )
             except UserNotParticipant:
                 if chat_id in links:
                     invitelink = links[chat_id]
@@ -142,7 +152,10 @@ def PlayWrapper(command):
                             pass
                     else:
                         try:
-                            invitelink = await app.export_chat_invite_link(chat_id)
+                            await client.get_chat_member(message.chat.id, "me")
+                            invitelink = await client.export_chat_invite_link(
+                                message.chat.id
+                            )
                         except ChatAdminRequired:
                             return await message.reply_text(_["call_1"])
                         except Exception as e:
@@ -154,7 +167,7 @@ def PlayWrapper(command):
                     invitelink = invitelink.replace(
                         "https://t.me/+", "https://t.me/joinchat/"
                     )
-                myu = await message.reply_text(_["call_4"].format(app.mention))
+                myu = await message.reply_text(_["call_5"])
                 try:
                     await asyncio.sleep(1)
                     await userbot.join_chat(invitelink)
@@ -162,19 +175,19 @@ def PlayWrapper(command):
                     try:
                         await app.approve_chat_join_request(chat_id, userbot.id)
                     except Exception as e:
-                        return await message.reply_text(
-                            _["call_3"].format(app.mention, type(e).__name__)
-                        )
-                    await asyncio.sleep(3)
-                    await myu.edit(_["call_5"].format(app.mention))
+                        return await myu.edit(_["call_3"].format(type(e).__name__))
+                    await asyncio.sleep(1)
+                    await myu.edit(_["call_6"].format(app.mention))
                 except UserAlreadyParticipant:
                     pass
                 except Exception as e:
-                    return await message.reply_text(
-                        _["call_3"].format(app.mention, type(e).__name__)
-                    )
+                    return await myu.edit(_["call_3"].format(type(e).__name__))
 
                 links[chat_id] = invitelink
+                try:
+                    await myu.delete()
+                except Exception:
+                    pass
 
                 try:
                     await userbot.resolve_peer(chat_id)
